refactor(runbotics-orchestrator-ui): :lipstick: Refactor for VariableMenu panel
