+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/ProcessVariablesPanel/VariableMenu.tsx
@@ -1,0 +1,27 @@
+import { FC } from 'react';
+
+import { Menu } from '@mui/material';
+
+import { VariableMenuProps, VariableCopyMenu } from './VariablesPanel';
+
+const VariableMenu: FC<VariableMenuProps> = ({ anchorElement, handleMenuClose, menuId }) => {
+    const isOpen = Boolean(menuId);
+
+    if (!anchorElement) {
+        return null;
+    }
+
+    return (
+        <Menu
+            key={menuId}
+            onClose={handleMenuClose}
+            id={menuId}
+            anchorEl={anchorElement}
+            open={isOpen}
+        >
+            <VariableCopyMenu menuId={menuId} handleMenuClose={handleMenuClose}/>
+        </Menu>
+    );
+};
+
+export default VariableMenu;
